Created RomGenerationService
