+++ src/Randomizer.SMZ3/Generation/GeneratedRomResult.cs
@@ -1,0 +1,9 @@
+﻿namespace Randomizer.SMZ3.Generation;
+using Randomizer.Shared.Models;
+
+public class GeneratedRomResult
+{
+    public GeneratedRom? Rom { get; set; }
+
+    public string? GenerationError { get; set; }
+}
