+++ src/Randomizer.SMZ3/Randomizer.SMZ3.csproj
@@ -44,6 +44,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Randomizer.Data\Randomizer.Data.csproj" />
     <ProjectReference Include="..\Randomizer.Shared\Randomizer.Shared.csproj" />
+    <ProjectReference Include="..\Randomizer.Sprites\Randomizer.Sprites.csproj" />
   </ItemGroup>
 
 </Project>
